requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
billiard==3.3.0.23
celery==3.1.26.post2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery==3.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
kombu==3.0.37
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
PyJWT==2.3.0
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
